Rename Minutely to Probes module
Make the minutely probes naming consistent between integrations, by
renaming the Minutely module to the Probes module.

Add a redirect for the Minutely constant to the Probes constant, so
people can upgrade without their app breaking. The deprecation warning
message that gets printed and logged when the Minutely module gets
called, will point to the location people need to change to remove the
warning.

Fixes #813
